coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2653 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/datadriven/algorithm/AlgorithmAdaBoostBase.cpp
@@ -969,7 +969,7 @@ namespace datadriven
 			}
 			else if (this->refineMode == 2)
 			{
-				refineNumber = (size_t)(this->perOfAda * (this->grid->getSize()));
+			  refineNumber = (size_t)(this->perOfAda * (double)(this->grid->getSize()));
 				//force to refine at least one point
 				if(refineNumber == 0)
 					refineNumber = 1;
