+++ src/sgpp/datadriven/basis/linear/boundary/operation/OperationRegularizationDiagonalLinearBoundary.cpp
@@ -21,13 +21,13 @@ namespace datadriven
     }
 
   void OperationRegularizationDiagonalLinearBoundary::initHkmix (double k) {
-    int dim = storage->dim();
+    size_t dim = storage->dim();
     base::GridIndex* gi;
     double res;
     for (size_t i=0; i<size; i++) {
       gi = storage->get(i);
       res = 1.0;
-      for (int d=0; d<dim; d++) {
+      for (size_t d=0; d<dim; d++) {
 	res *= pow(2, (2*k-1) * gi->getLevel(d) -1);
       }
       diagonal[i] = res;
@@ -36,20 +36,20 @@ namespace datadriven
   }
   
   void OperationRegularizationDiagonalLinearBoundary::initH0HkLaplace (double k) {
-    int dim = storage->dim();
+    size_t dim = storage->dim();
     base::GridIndex* gi;
     double res, resd;
     for (size_t i=0; i<size; i++) {
       gi = storage->get(i);
       res = 0.0;
-      for (int d=0; d<dim; d++) {
+      for (size_t d=0; d<dim; d++) {
 	// Hk in dimension d
 	resd = pow(2, (2*k-1) * gi->getLevel(d) -1);
 	// "H0" in remaining dimensions
-	for (int d2=0; d2<d; d2++) {
+	for (size_t d2=0; d2<d; d2++) {
 	  resd *= pow(2.0, -1 -gi->getLevel(d2));
 	}
-	for (int d2=d+1; d2<dim; d2++) {
+	for (size_t d2=d+1; d2<dim; d2++) {
 	  resd *= pow(2.0, -1 -gi->getLevel(d2));
 	}
 	res += resd;
